logger/requirements.txt
nameko==2.14.0
nameko-sqlalchemy==1.5.0
alembic==1.0.10
marshmallow==2.19.2
psycopg2-binary==2.8.2
pytest==4.5.0
coverage==4.5.3
flake8==4.0.1
flake8-annotations==2.7.0